requirements.txt
fastapi>=0.115.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
openai>=1.40.0
litert-lm-api-nightly
